Updated framework with browser code and page object model identification

+++ src/test/java/testScripts/CreateAccount.java
@@ -1,32 +1,53 @@
+/*
+ * This is a basic test case to create account on Amazon
+ * */
 package testScripts;
-
-import java.util.Random;
-
 import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.xerces.util.URI.MalformedURIException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
+
+import InvokeBrowser.InvokeBrowserSettings;
+
 import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.io.TemporaryFilesystem;
 
 
 
 public class CreateAccount {
 	
-	//WebDriver driver;
-	@Test
-	public void TC_01() {
+	public WebDriver driver;
+	
+	@BeforeMethod
+	@Parameters({"URL"})
+	public void browserSetup(String URL) throws MalformedURIException{
 		System.setProperty("webdriver.chrome.driver","//Users//anjalimutha//Downloads//chromedriver");
-		String baseUrl = "https://www.amazon.com/";	
+		InvokeBrowserSettings invokebrwsr = new InvokeBrowserSettings();
+		invokebrwsr.invokeBrowser("Chrome", URL);
+	}
+	
+	@AfterMethod
+	public void cleanUp() {
+		if(driver!=null) {
+			TemporaryFilesystem.getDefaultTmpFS().deleteTemporaryFiles();
+			driver.manage().deleteAllCookies();
+			driver.close();
+			driver.quit();
+		}
+	}
+
+	
+	@Test
+	public void TC_01_CreateAccount() {
 		
-		WebDriver driver = new ChromeDriver();
-		driver.manage().deleteAllCookies();
-		driver.get(baseUrl);
 		String cust = RandomStringUtils.randomAlphabetic(10);
 		String email=RandomStringUtils.randomAlphanumeric(8) + "@gmail.com";
 		
-		driver.manage().window().maximize();
 		Actions action = new Actions(driver);
 		WebElement element= driver.findElement(By.xpath("//span[@class='nav-line-2'][contains(text(),'Account & Lists')]"));
 		action.moveToElement(element).build().perform();
@@ -39,11 +60,7 @@ public class CreateAccount {
 		
 		//driver.findElement(By.xpath("//div[@id='hud-customer-name']/div/a[contains(text(),'\"+cust+\"')]")).isDisplayed();
 		
-		driver.close();
-				
+		
 	}
 	
-	
-	
-
 }
